database: move BasicRegistry class as a shared base class to the utils module
The base class has some nice tools that we want to have available for other
classes as well, so it makes sense to move it to the utils module.

utils: implement multi threading for EventSource

We want to have the EventSource running in the background on most of the
live reading activities. This change allows to have it running in background.